Rewrite the way we read in files.
* Move file reading into Input rather than repl.
* Handle file command in run in repl.
* Create unexecuted_commands list containing all of the commands left to execute and handle them inside run.
* Move command line args into unexecuted_commands.
