+++ tools/cli/src/mentat_cli/input.rs
@@ -13,7 +13,14 @@ use std::io::stdin;
 use linefeed::Reader;
 use linefeed::terminal::DefaultTerminal;
 
+use std::io::BufReader;
+use std::io::BufRead;
+use std::fs::File;
+use std::path::Path;
+
 use self::InputResult::*;
+
+use edn;
 
 use command_parser::{
     Command, 
@@ -69,6 +76,44 @@ impl InputReader {
     /// Returns whether the `InputReader` is reading from a TTY.
     pub fn is_tty(&self) -> bool {
         self.reader.is_some()
+    }
+
+    pub fn read_file(&mut self, filename: &String) -> Result<Vec<InputResult>, cli::Error> {
+        let path = Path::new(&filename);
+        let display = path.display();
+
+        let f = match File::open(&path) {
+            Err(err) => bail!(cli::ErrorKind::FileError(display.to_string(), err.to_string())),
+            Ok(file) => file,
+        };
+
+        let mut buffer = String::new();
+
+        let file = BufReader::new(&f);
+        let mut cmd_err: Option<edn::ParseError> = None;
+
+        let mut results: Vec<InputResult> = Vec::new();
+        for line in file.lines() {
+            let l = line.unwrap();
+            println!("{}", l);
+            buffer.push_str(&l);
+            let cmd = Command::Transact(buffer.to_string());
+            let (is_complete, err) = cmd.is_complete();
+            if is_complete {
+                results.push(InputResult::MetaCommand(cmd));
+                buffer.clear();
+                cmd_err = None;
+            } else {
+                cmd_err = err;
+            }
+
+        }
+
+        if let Some(err) = cmd_err {
+            bail!(err);
+        }
+
+        Ok(results)
     }
 
     /// Reads a single command, item, or statement from `stdin`.
